events system (#55)
add events system to replace current "realtime" setup

current realtime setup is a bit dodgy - we notify the client when
relevant data has updated, and then refetch the data in a separate
request from the client. this was required as some queries involved
joins, and you can't subscribe to "join"-ed queries with supabase's
realtime. that was good enough to get the game to work, but makes the
game feel a bit janky, and slower to respond than is ideal.

the events system will allow us to easily send fat events containing the
updated data for the client. for now, it is built on top of a postgres
table that we listen to w/ supabase's realtime, but should be relatively
easily swappable for our own or a 3rd party websocket service. for now,
i've pushed & handle messages for joining games as a POC, but should be
easy to add new events & remove the old realtime-ish setup.

old events (>1 week old) will be deleted. picked 1 week so we have a
chance to debug issues

Ref: CHA-185